Add a special clog representation for Sequel::Model
We redact encrypted and lengthy columns from the .inspect output in our
models. It would be wisely to apply the same redaction to our clog
metadata as well.

Also, if the `Sequel::Model` is provided as metadata, the model
automatically determines the key name from its table name and returns it
with redacted values. `{ strand: strand.values }` can be abbreviated as
`{ strand }`.