Mark virtual machines as tainted
At a certain stage in a virtual machine's lifecycle, it may become
tainted with user data.

From the moment we start the hypervisor for the first time, it's
uncertain when the user logged in and began processing. After this, the
virtual machine is deemed tainted if it wasn't created for a VM pool.
Virtual machines from the pool are considered tainted once they are
selected from the pool. Customers cannot access them until they are
chosen from the pool. If the virtual machine hasn't been marked as
tainted before its destroy, we mark it as tainted during the destroy
process.

When a virtual machine is considered tainted, it theoretically contains
irreplaceable and private data.

We can develop various logics based on the taint value.

For instance, we discussed some 'recreate' logics for virtual machines.
We can easily recreate untainted virtual machines, but we need to
exercise caution with tainted ones.

Furthermore, the taint value can be utilized in our pool logic. We can
select only untainted virtual machines from the pool.