Use provisioned at instead of display state to pick a vm from the pool
The 'display_state' function is used to show the current state of the
virtual machines to the customer, and it's dynamically calculated in
certain cases. The 'display' prefix implies that we shouldn't use it
into our backend logics. 'provisioned_at' serves as a more reliable
indicator to determine whether the virtual machine has been provisioned.